grpc-ex/pyproject.toml
[tool.poetry]
name = "grpc-ex"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
grpcio = "^1.24"
grpcio-tools = "^1.24"
pysqlite3 = "^0.4.1"
pandas = "^0.25.2"
PyInquirer = "^1.0"

[tool.poetry.dev-dependencies]
ipdb = "^0.12.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
